Share bug fixing
1) Save image files (for sharing purpose) in external but application
specific directory since this is not typically visible to the user as
media and these files will be deleted when the application is
uninstalled.
2) On each restart of app process clear the above storage.
3) Don't overwrite image files because we can have more than 1 screens
in back stack which has sharing functionality. Hence create unique image
name based on timestamp & then delete all these on app restart as
mentioned under point 2.
4) If venue details screen is loaded & its image loading is in progress,
then even if user attempts to share this venue after image loading is
completed, this image is not attached with venue sharing email. This
happens because on finishing image loading, notification is not sent to
update share intent. Now this is fixed.