Bug fixing & updates
1) Bug fixing:
a) change location - zoom is too far in - should show the city you are
currently viewing.
b) event details - the description should load before the map does on
the info tab.
2) On continue button click from fb login screen redirect user to
connect accounts screen.
3) On continue button click from connect accounts screen, there must be
a conditional redirection to my events or discover screen depending on
whether any my events are available or not.