Bug fixing
a) If we start artists news screen in horizontal mode it was not showing
full-width images.
b) If application is not launched & user gets more than 1 notifications
then clicking on any one notification redirects to event details screen,
but after this clicking on any of the remaining notifications does not
refresh event details screen.
c) If no artists are found the '+' symbol should be replaced with a '0'
to show that it's searched.
d) Widget should only show the events for the city the app is currently
connected to.  I had events for san francisco and chicago.
e) Widget - arrows are too hard to push - might want to make them a bit
larger.
f) widget - when image has not loaded please remove the eventseeker logo
background or make it fit the space correctly.