Friends feed bug fixing
Under my friends feed there was an event whose schedule and venue
details were not available. When queried for this specific event its
record was not found on server. On client side it used to prevent
fetching of further feeds which is solved now. It will skip displaying
such invalid events on friends feed screen though it's found in
response.