Bug fixing, inception of Proguard usage
1) Under my friends feed there was an event whose schedule and venue
details were not available. When queried for this specific event its
record was not found on server. On client side it used to prevent
fetching of further feeds which is solved now. It will skip displaying
such invalid events on friends feed screen though it's found in
response.
2) If phone number for any venue is unavailable, don't start any action
when user presses phone button on venue info/events screen.
3) When clicked on tickets button for an event named 'Larry Chiang Film
Pr Parties Back-to-back-to-back Before...', it opens up an empty screen.
This is fixed by enabling javascript on our webview.
4) Proguard usage inception.