removed launchmode=singletask due to error in Bosch
removed launchmode=singletask due to error in BoschMainActivity.

Error: In Bosch App on pressing 'Home' button app gets into background
and when pressed app launcher it should resume its previous state. But
due to launchmode=singletask attribute in MainActivity, the
BoachMainActivity(launched from MainActivity) was getting destroyed, on
pressing app launcher icon and was resuming the MainActivity and from
that it was discovering the connection with Bosch System and was
starting new BoschMainActivity and a new session.