search view, blue event slider, hamburger bg, etc.
1) When clicking the side navigation button it should not show a light
box around the icon.
2) Event details -
a) Should show floating action buttons when sitting at bottom of page.
b) On the featuring overlay the word is sitting too high from the
bottom. It should fit more near middle/bottom of the overlay.
3) On 4 screens - Discover, Artist details, Venue details, Search
events, when clicking the More button for the slide the dots should turn
blue when pressing and not show a box around the icon.
4) The line to type what you are searching for needs to be white, it was
green until now.
5) Removed redundant file searchview.xml