Signup screen, keypad state, hamburger bug fix
1) Login & sign up screens - windowSoftInputMode="stateUnchanged" is
used to retain soft keypad status (open/close) on orientation change &
"adjustResize" is to allow access to entire layout when keypad is open &
to keep toolbar as it is which used to scroll up (becoming invisible) in
absence of this "adjustResize" value.
2) Add top padding for tablet main content frame in
activity_base_tab.xml
3) Sign up & login screens for tablet.
4) 10" tablet toolbar title text size increased.
5) Bug fix: On 10" tablet isTaskRoot() was not returning right value
instantly hence it was setting back arrow instead of hamburger icon
sometimes.
6) Don't execute bosch & ford related code for tablet  login & sign up
screens.