Bug fix - parse cat id, etc; discover memory mgt
1) Discover bug fixing - For builds prepared from android studio, on
orientation change category selection was changing. Fixed this by using
handler for onscroll listener in discover tab screen so that onglobal
layout gets called before on scroll listener.
2) Memory handling on discover screen - Remove 'no content' background
images when not needed.
3) Artist details - Parse category 'id' in addition to cat, because in
some cases 'cat' key doesn't come, based on which we are deciding if we
need to load upcoming events or not (for sports cat).